// packages/comments/src/generated/graphqlgen.ts
export namespace QueryResolvers {
    export const defaultResolvers = {}

    export interface ArgsQueryComment {
        parentId: string
        limit: number | null
        offset: number | null
    }

    export type QueryCommentResolver = (
        parent: undefined,
        args: ArgsQueryComment,
        ctx: Context,
        info: GraphQLResolveInfo,
    ) => CommentConnection | Promise<CommentConnection>

    export interface Type {
        queryComment: (
            parent: undefined,
            args: ArgsQueryComment,
            ctx: Context,
            info: GraphQLResolveInfo,
        ) => CommentConnection | Promise<CommentConnection>
    }
}
// ...
    export const defaultResolvers = {
        pageInfo: (parent: CommentConnection) => parent.pageInfo,
        edges: (parent: CommentConnection) => parent.edges,
    }
// ...
export namespace MutationResolvers {
    export const defaultResolvers = {}

    export interface ArgsDeleteComment {
        id: string
    }

    export interface ArgsEditComment {
        id: string
        body: string
    }

    export interface ArgsCreateComment {
        pageId: string
        parentId: string
        body: string | null
        repliedTo: string | null
    }

    export interface ArgsCreateReply {
        pageId: string
        parentId: string
        body: string | null
        repliedTo: string | null
    }

    export interface ArgsLikeComment {
        commentId: string
    }

    export type DeleteCommentResolver = (
        parent: undefined,
        args: ArgsDeleteComment,
        ctx: Context,
        info: GraphQLResolveInfo,
    ) => DeleteCommentResponse | Promise<DeleteCommentResponse>

    export type EditCommentResolver = (
        parent: undefined,
        args: ArgsEditComment,
        ctx: Context,
        info: GraphQLResolveInfo,
    ) => Comment | Promise<Comment>

    export type CreateCommentResolver = (
        parent: undefined,
        args: ArgsCreateComment,
        ctx: Context,
        info: GraphQLResolveInfo,
    ) => Comment | Promise<Comment>

    export type CreateReplyResolver = (
        parent: undefined,
        args: ArgsCreateReply,
        ctx: Context,
        info: GraphQLResolveInfo,
    ) => Comment | Promise<Comment>

    export type LikeCommentResolver = (
        parent: undefined,
        args: ArgsLikeComment,
        ctx: Context,
        info: GraphQLResolveInfo,
    ) => Comment | null | Promise<Comment | null>

    export interface Type {
        deleteComment: (
            parent: undefined,
            args: ArgsDeleteComment,
            ctx: Context,
            info: GraphQLResolveInfo,
        ) => DeleteCommentResponse | Promise<DeleteCommentResponse>

        editComment: (
            parent: undefined,
            args: ArgsEditComment,
            ctx: Context,
            info: GraphQLResolveInfo,
        ) => Comment | Promise<Comment>

        createComment: (
            parent: undefined,
            args: ArgsCreateComment,
            ctx: Context,
            info: GraphQLResolveInfo,
        ) => Comment | Promise<Comment>

        createReply: (
            parent: undefined,
            args: ArgsCreateReply,
            ctx: Context,
            info: GraphQLResolveInfo,
        ) => Comment | Promise<Comment>

        likeComment: (
            parent: undefined,
            args: ArgsLikeComment,
            ctx: Context,
            info: GraphQLResolveInfo,
        ) => Comment | null | Promise<Comment | null>
    }
}
// ...
export namespace SubscriptionResolvers {
    export const defaultResolvers = {}

    export interface ArgsNewCommentSubscription {
        pageId: string
    }

    export type NewCommentSubscriptionResolver = {
        subscribe: (
            parent: undefined,
            args: ArgsNewCommentSubscription,
            ctx: Context,
            info: GraphQLResolveInfo,
        ) =>
            | AsyncIterator<CommentSubscriptionPayload>
            | Promise<AsyncIterator<CommentSubscriptionPayload>>
        resolve?: (
            parent: undefined,
            args: ArgsNewCommentSubscription,
            ctx: Context,
            info: GraphQLResolveInfo,
        ) => CommentSubscriptionPayload | Promise<CommentSubscriptionPayload>
    }

    export interface Type {
        newCommentSubscription: {
            subscribe: (
                parent: undefined,
                args: ArgsNewCommentSubscription,
                ctx: Context,
                info: GraphQLResolveInfo,
            ) =>
                | AsyncIterator<CommentSubscriptionPayload>
                | Promise<AsyncIterator<CommentSubscriptionPayload>>
            resolve?: (
                parent: undefined,
                args: ArgsNewCommentSubscription,
                ctx: Context,
                info: GraphQLResolveInfo,
            ) =>
                | CommentSubscriptionPayload
                | Promise<CommentSubscriptionPayload>
        }
    }
}